+++ shared/src/metrics.rs
@@ -1,6 +1,6 @@
 use {
     crate::config::ConfigMetrics,
-    http_body_util::{combinators::BoxBody, BodyExt, Empty as BodyEmpty, Full as BodyFull},
+    http_body_util::{BodyExt, Full as BodyFull},
     hyper::{
         body::{Bytes, Incoming as BodyIncoming},
         service::service_fn,
@@ -11,7 +11,7 @@ use {
         server::conn::auto::Builder as ServerBuilder,
     },
     prometheus::{proto::MetricFamily, TextEncoder},
-    std::{convert::Infallible, future::Future},
+    std::future::Future,
     tokio::{net::TcpListener, task::JoinError},
     tracing::{error, info},
 };
@@ -19,6 +19,8 @@ use {
 pub async fn spawn_server(
     ConfigMetrics { endpoint }: ConfigMetrics,
     gather_metrics: impl Fn() -> Vec<MetricFamily> + Clone + Send + 'static,
+    is_health_check: impl Fn() -> bool + Clone + Send + 'static,
+    is_ready_check: impl Fn() -> bool + Clone + Send + 'static,
     shutdown: impl Future<Output = ()> + Send + 'static,
 ) -> std::io::Result<impl Future<Output = Result<(), JoinError>>> {
     let listener = TcpListener::bind(endpoint).await?;
@@ -43,17 +45,56 @@ pub async fn spawn_server(
                 },
             };
             let gather_metrics = gather_metrics.clone();
+            let is_health_check = is_health_check.clone();
+            let is_ready_check = is_ready_check.clone();
             tokio::spawn(async move {
                 if let Err(error) = ServerBuilder::new(TokioExecutor::new())
                     .serve_connection(
                         TokioIo::new(stream),
                         service_fn(move |req: Request<BodyIncoming>| {
                             let gather_metrics = gather_metrics.clone();
+                            let is_health_check = is_health_check.clone();
+                            let is_ready_check = is_ready_check.clone();
                             async move {
-                                match req.uri().path() {
-                                    "/metrics" => metrics_handler(&gather_metrics()),
-                                    _ => not_found_handler(),
-                                }
+                                let (status, bytes) = match req.uri().path() {
+                                    "/health" => {
+                                        if is_health_check() {
+                                            (StatusCode::OK, Bytes::from("OK"))
+                                        } else {
+                                            (
+                                                StatusCode::INTERNAL_SERVER_ERROR,
+                                                Bytes::from("Service is unhealthy"),
+                                            )
+                                        }
+                                    }
+                                    "/metrics" => {
+                                        let metrics = TextEncoder::new()
+                                            .encode_to_string(&gather_metrics())
+                                            .unwrap_or_else(|error| {
+                                                error!(
+                                                    "could not encode custom metrics: {}",
+                                                    error
+                                                );
+                                                String::new()
+                                            });
+                                        (StatusCode::OK, Bytes::from(metrics))
+                                    }
+                                    "/ready" => {
+                                        if is_ready_check() {
+                                            (StatusCode::OK, Bytes::from("OK"))
+                                        } else {
+                                            (
+                                                StatusCode::INTERNAL_SERVER_ERROR,
+                                                Bytes::from("Service is not ready"),
+                                            )
+                                        }
+                                    }
+                                    _ => (StatusCode::NOT_FOUND, Bytes::new()),
+                                };
+
+                                Response::builder()
+                                    .status(status)
+                                    .body(BodyFull::new(bytes).boxed())
                             }
                         }),
                     )
@@ -64,24 +105,4 @@ pub async fn spawn_server(
             });
         }
     }))
-}
-
-fn metrics_handler(
-    metric_families: &[MetricFamily],
-) -> http::Result<Response<BoxBody<Bytes, Infallible>>> {
-    let metrics = TextEncoder::new()
-        .encode_to_string(metric_families)
-        .unwrap_or_else(|error| {
-            error!("could not encode custom metrics: {}", error);
-            String::new()
-        });
-    Response::builder()
-        .status(StatusCode::OK)
-        .body(BodyFull::new(Bytes::from(metrics)).boxed())
-}
-
-fn not_found_handler() -> http::Result<Response<BoxBody<Bytes, Infallible>>> {
-    Response::builder()
-        .status(StatusCode::NOT_FOUND)
-        .body(BodyEmpty::new().boxed())
 }